refactor(input): remove inner state of input & new onchange propagation
affects: bl-input
