+++ packages/input/index.ts
@@ -1,1 +1,5 @@
-import './Input';
+import { define } from 'skatejs';
+import { Input } from './Input';
+
+export { Input } from './Input';
+define( Input );
